+++ websites/F/Fire Emblem Heroes/pages/newguide.ts
@@ -1,0 +1,5 @@
+export default function applyNewGuideDetails(presenceData: PresenceData): void {
+	presenceData.details = "Browsing the Guide";
+	presenceData.state = document.querySelector("h1").textContent;
+	presenceData.buttons = [{ label: "Read Guide", url: document.location.href }];
+}
